♻ refactor: score 타입 변경 double -> long

+++ src/main/java/com/example/peachmusic/domain/ranking/model/RankingResponseDto.java
@@ -9,13 +9,13 @@ import org.springframework.data.redis.core.ZSetOperations.TypedTuple;
 public class RankingResponseDto {
 
     private final String title;
-    private final double score;
+    private final Long score;
 
     // Tuple을 사용하여 RankingResponseDto를 생성하는 메서드
     public static RankingResponseDto of(TypedTuple<String> tuple) {
         return new RankingResponseDto(
                 tuple.getValue(),
-                tuple.getScore() // score는 double 타입
+                tuple.getScore().longValue()
         );
     }
 }
